Extract user overlay actions button component
