+++ osu.Game/Overlays/Profile/Header/Components/UserActionsButton.cs
@@ -1,7 +1,6 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System.Collections.Generic;
 using System.Linq;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
@@ -9,34 +8,20 @@ using osu.Framework.Extensions;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Cursor;
-using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
 using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Input.Events;
-using osu.Framework.Localisation;
-using osu.Game.Graphics;
-using osu.Game.Graphics.Containers;
-using osu.Game.Graphics.Sprites;
-using osu.Game.Graphics.UserInterfaceV2;
 using osu.Game.Online.API;
 using osu.Game.Online.API.Requests.Responses;
 using osu.Game.Resources.Localisation.Web;
 using osu.Game.Users;
-using osuTK;
 
 namespace osu.Game.Overlays.Profile.Header.Components
 {
-    public partial class UserActionsButton : OsuHoverContainer, IHasPopover
+    public partial class UserActionsButton : ProfileActionsButton
     {
         public readonly Bindable<UserProfileData?> User = new Bindable<UserProfileData?>();
 
-        private Box background = null!;
         private UserReportPopoverTarget reportPopoverTarget = null!;
-
-        protected override IEnumerable<Drawable> EffectTargets => [background];
-
-        [Resolved]
-        private OverlayColourProvider colourProvider { get; set; } = null!;
 
         [Resolved]
         private IAPIProvider api { get; set; } = null!;
@@ -44,45 +29,18 @@ namespace osu.Game.Overlays.Profile.Header.Components
         [BackgroundDependencyLoader]
         private void load()
         {
-            IdleColour = colourProvider.Background2;
-            HoverColour = colourProvider.Background1;
-
-            Size = new Vector2(40);
-            Masking = true;
-            CornerRadius = 20;
-
-            Child = new CircularContainer
+            // This is a bit of a dirty hack. Because `ReportUserPopover` is spawned from `UserActionsPopover`,
+            // and that they both share the same `PopoverContainer`, the former will get destroyed when the latter
+            // is opened, causing it to get destroyed as well.
+            //
+            // This is worked around by having an additional dummy popover target on the actions button,
+            // which is then passed to `UserActionsPopover` and the user report action. This way the popover
+            // can remain attached to it once the actions popover is destroyed.
+            reportPopoverTarget = new UserReportPopoverTarget
             {
-                Masking = true,
-                RelativeSizeAxes = Axes.Both,
-                Children = new Drawable[]
-                {
-                    background = new Box
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                    },
-                    // This is a bit of a dirty hack. Because `ReportUserPopover` is spawned from `UserActionsPopover`,
-                    // and that they both share the same `PopoverContainer`, the former will get destroyed when the latter
-                    // is opened, causing it to get destroyed as well.
-                    //
-                    // This is worked around by having an additional dummy popover target on the actions button,
-                    // which is then passed to `UserActionsPopover` and the user report action. This way the popover
-                    // can remain attached to it once the actions popover is destroyed.
-                    reportPopoverTarget = new UserReportPopoverTarget
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                    },
-                    new SpriteIcon
-                    {
-                        Size = new Vector2(12),
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        Icon = FontAwesome.Solid.EllipsisV,
-                    },
-                }
+                RelativeSizeAxes = Axes.Both
             };
-
-            Action = this.ShowPopover;
+            Add(reportPopoverTarget);
         }
 
         protected override void LoadComplete()
@@ -96,142 +54,44 @@ namespace osu.Game.Overlays.Profile.Header.Components
             }, true);
         }
 
-        public Popover GetPopover() => new UserActionPopover(User.Value!.User, reportPopoverTarget);
+        public override Popover GetPopover() => new UserActionPopover(User.Value!.User, reportPopoverTarget);
 
-        private partial class UserActionPopover : OsuPopover
+        private partial class UserActionPopover : ProfileActionPopover
         {
             private readonly APIUser user;
 
             private readonly IHasPopover reportPopoverTarget;
 
             public UserActionPopover(APIUser user, IHasPopover reportPopoverTarget)
-                : base(false)
             {
                 this.user = user;
                 this.reportPopoverTarget = reportPopoverTarget;
             }
 
             [BackgroundDependencyLoader]
-            private void load(OverlayColourProvider colourProvider, IAPIProvider api, IDialogOverlay? dialogOverlay)
+            private void load(IAPIProvider api, IDialogOverlay? dialogOverlay)
             {
-                Background.Colour = colourProvider.Background6;
-
                 bool userBlocked = api.LocalUserState.Blocks.Any(b => b.TargetID == user.Id);
 
-                AllowableAnchors = [Anchor.BottomCentre, Anchor.TopCentre];
-
-                Child = new FillFlowContainer
+                Actions = new[]
                 {
-                    Width = 160,
-                    AutoSizeAxes = Axes.Y,
-                    Padding = new MarginPadding { Horizontal = 5, Vertical = 10 },
-                    Children = new Drawable[]
+                    new ProfilePopoverAction(FontAwesome.Solid.Ban, userBlocked ? UsersStrings.BlocksButtonUnblock : UsersStrings.BlocksButtonBlock)
                     {
-                        new UserAction(FontAwesome.Solid.Ban, userBlocked ? UsersStrings.BlocksButtonUnblock : UsersStrings.BlocksButtonBlock)
+                        Action = () =>
                         {
-                            Action = () =>
-                            {
-                                dialogOverlay?.Push(userBlocked ? ConfirmBlockActionDialog.Unblock(user) : ConfirmBlockActionDialog.Block(user));
-                                this.HidePopover();
-                            }
-                        },
-                        new UserAction(FontAwesome.Solid.ExclamationTriangle, ReportStrings.UserButton)
-                        {
-                            Action = () =>
-                            {
-                                this.HidePopover();
-                                reportPopoverTarget.ShowPopover();
-                            },
+                            dialogOverlay?.Push(userBlocked ? ConfirmBlockActionDialog.Unblock(user) : ConfirmBlockActionDialog.Block(user));
+                            this.HidePopover();
                         }
-                    }
-                };
-            }
-        }
-
-        private partial class UserAction : OsuClickableContainer
-        {
-            private readonly IconUsage icon;
-            private readonly LocalisableString caption;
-
-            private Box background = null!;
-            private CircularContainer indicator = null!;
-
-            public UserAction(IconUsage icon, LocalisableString caption)
-            {
-                this.icon = icon;
-                this.caption = caption;
-            }
-
-            [BackgroundDependencyLoader]
-            private void load(OverlayColourProvider colourProvider)
-            {
-                RelativeSizeAxes = Content.RelativeSizeAxes = Axes.X;
-                AutoSizeAxes = Content.AutoSizeAxes = Axes.Y;
-
-                Masking = true;
-                CornerRadius = 4;
-                Children = new Drawable[]
-                {
-                    background = new Box
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                        Colour = colourProvider.Background5,
-                        Alpha = 0,
                     },
-                    indicator = new Circle
+                    new ProfilePopoverAction(FontAwesome.Solid.ExclamationTriangle, ReportStrings.UserButton)
                     {
-                        Width = 4,
-                        Height = 14,
-                        X = 10,
-                        Colour = colourProvider.Highlight1,
-                        Anchor = Anchor.CentreLeft,
-                        Origin = Anchor.Centre,
-                        Alpha = 0,
-                    },
-                    new FillFlowContainer
-                    {
-                        AutoSizeAxes = Axes.Y,
-                        RelativeSizeAxes = Axes.X,
-                        Padding = new MarginPadding { Horizontal = 25, Vertical = 5 },
-                        Direction = FillDirection.Horizontal,
-                        Spacing = new Vector2(5, 0),
-                        Children = new Drawable[]
+                        Action = () =>
                         {
-                            new SpriteIcon
-                            {
-                                Icon = icon,
-                                Size = new Vector2(11),
-                                Anchor = Anchor.CentreLeft,
-                                Origin = Anchor.CentreLeft,
-                            },
-                            new OsuSpriteText
-                            {
-                                Text = caption,
-                                Font = OsuFont.Style.Caption1,
-                                Anchor = Anchor.CentreLeft,
-                                Origin = Anchor.CentreLeft,
-                                UseFullGlyphHeight = false,
-                            }
+                            this.HidePopover();
+                            reportPopoverTarget.ShowPopover();
                         }
-                    }
+                    },
                 };
-            }
-
-            protected override bool OnHover(HoverEvent e)
-            {
-                updateState();
-                return true;
-            }
-
-            protected override void OnHoverLost(HoverLostEvent e)
-            {
-                updateState();
-                base.OnHoverLost(e);
-            }
-
-            private void updateState()
-            {
-                background.Alpha = indicator.Alpha = IsHovered ? 1 : 0;
             }
         }
